Replaced BotTypeContext with built-in AnnotatedElementModifier
